+++ app/javascript/vue/components/Filter/Facets/shared/FacetDataAttribute/AddPredicate.vue
@@ -60,7 +60,7 @@
           <VBtn
             color="primary"
             medium
-            :disabled="!predicate"
+            :disabled="!predicate || !inputValue.length"
             @click="() => addPredicate({ any: false })"
           >
             Add
@@ -68,20 +68,12 @@
           <VBtn
             color="primary"
             medium
-            :disabled="!predicate || !!inputValue"
+            :disabled="!predicate || !!inputValue.length"
             @click="() => addPredicate({ any: true, text: '', exact: false })"
           >
             Any
           </VBtn>
         </div>
-        <label>
-          <input
-            v-model="exact"
-            :disabled="!inputValue"
-            type="checkbox"
-          />
-          Exact
-        </label>
       </div>
     </div>
   </div>
@@ -100,7 +92,6 @@ const emit = defineEmits(['add'])
 const smartSelectorRef = ref(null)
 const inputValue = ref('')
 const predicate = ref()
-const exact = ref(false)
 const list = ref([])
 
 function addPredicate(params) {
@@ -108,7 +99,7 @@ function addPredicate(params) {
     id: predicate.value.id,
     name: predicate.value.name,
     text: inputValue.value,
-    exact: exact.value,
+    exact: false,
     ...params
   }